change api to answer on "powerdns" server-id for internal calls. Adjusted Tests.
Partially reverting back the previous change to get the tests working with a server_id of 'powerdns' corresponding the other tests. I have not seen any usage of this api for powerdns-admin internal zones. Anyway this is now backwards compatible if you use the powerdns id instead of localhost. And giving the server_id "localhost" forwards the request to the real powerdns server responding with the correct answer.
Answering on this api endpoint with the data from powerdns-admin database is still wrong.