Update Dockerfile - fix to pip upgrade
The `pip install --upgrade pip` should be done first in its own RUN. This allows the subsequent `pip install -r` to use the updated `pip`.  If done in the same `RUN` the old version of pip loaded in the environment is used which causes problems for things like `cryptography` requiring rust compilers and such.