strip() whitespace from new user master data
When creating a new user, there is a chance that, due to a typing error,
whitespace will be introduced at the start or end of the username. This
can lead to issues when trying to log in using the affected username, as
such a condition can easily be overlooked - no user will be found in the
database if entering the username without the aforementioned whitespace.
This commit therefore strip()s the username string within
routes/{admin,index}.py.

The firstname, lastname and email strings within
routes/{admin,index,user}.py are also strip()ped on this occasion.